+++ lib/coinbase/client/api/smart_contracts_api.rb
@@ -328,5 +328,85 @@ module Coinbase::Client
       end
       return data, status_code, headers
     end
+
+    # Read data from a smart contract
+    # Perform a read operation on a smart contract without creating a transaction
+    # @param network_id [String] 
+    # @param contract_address [String] 
+    # @param read_contract_request [ReadContractRequest] 
+    # @param [Hash] opts the optional parameters
+    # @return [SolidityValue]
+    def read_contract(network_id, contract_address, read_contract_request, opts = {})
+      data, _status_code, _headers = read_contract_with_http_info(network_id, contract_address, read_contract_request, opts)
+      data
+    end
+
+    # Read data from a smart contract
+    # Perform a read operation on a smart contract without creating a transaction
+    # @param network_id [String] 
+    # @param contract_address [String] 
+    # @param read_contract_request [ReadContractRequest] 
+    # @param [Hash] opts the optional parameters
+    # @return [Array<(SolidityValue, Integer, Hash)>] SolidityValue data, response status code and response headers
+    def read_contract_with_http_info(network_id, contract_address, read_contract_request, opts = {})
+      if @api_client.config.debugging
+        @api_client.config.logger.debug 'Calling API: SmartContractsApi.read_contract ...'
+      end
+      # verify the required parameter 'network_id' is set
+      if @api_client.config.client_side_validation && network_id.nil?
+        fail ArgumentError, "Missing the required parameter 'network_id' when calling SmartContractsApi.read_contract"
+      end
+      # verify the required parameter 'contract_address' is set
+      if @api_client.config.client_side_validation && contract_address.nil?
+        fail ArgumentError, "Missing the required parameter 'contract_address' when calling SmartContractsApi.read_contract"
+      end
+      # verify the required parameter 'read_contract_request' is set
+      if @api_client.config.client_side_validation && read_contract_request.nil?
+        fail ArgumentError, "Missing the required parameter 'read_contract_request' when calling SmartContractsApi.read_contract"
+      end
+      # resource path
+      local_var_path = '/v1/networks/{network_id}/smart_contracts/{contract_address}/read'.sub('{' + 'network_id' + '}', CGI.escape(network_id.to_s)).sub('{' + 'contract_address' + '}', CGI.escape(contract_address.to_s))
+
+      # query parameters
+      query_params = opts[:query_params] || {}
+
+      # header parameters
+      header_params = opts[:header_params] || {}
+      # HTTP header 'Accept' (if needed)
+      header_params['Accept'] = @api_client.select_header_accept(['application/json']) unless header_params['Accept']
+      # HTTP header 'Content-Type'
+      content_type = @api_client.select_header_content_type(['application/json'])
+      if !content_type.nil?
+          header_params['Content-Type'] = content_type
+      end
+
+      # form parameters
+      form_params = opts[:form_params] || {}
+
+      # http body (model)
+      post_body = opts[:debug_body] || @api_client.object_to_http_body(read_contract_request)
+
+      # return_type
+      return_type = opts[:debug_return_type] || 'SolidityValue'
+
+      # auth_names
+      auth_names = opts[:debug_auth_names] || []
+
+      new_options = opts.merge(
+        :operation => :"SmartContractsApi.read_contract",
+        :header_params => header_params,
+        :query_params => query_params,
+        :form_params => form_params,
+        :body => post_body,
+        :auth_names => auth_names,
+        :return_type => return_type
+      )
+
+      data, status_code, headers = @api_client.call_api(:POST, local_var_path, new_options)
+      if @api_client.config.debugging
+        @api_client.config.logger.debug "API called: SmartContractsApi#read_contract\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
+      end
+      return data, status_code, headers
+    end
   end
 end