+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Include ERC20 and ERC721 token transfer information into address.transactions return value.
+
 ## [0.7.0] - 2024-09-26
 
 ### Added